+++ src/nftSDKFunctions/estimate-create-collection-in-dollars.ts
@@ -21,7 +21,7 @@ import { validatePropsForCreateCollection } from '../utils/validate-props';
 import { tokenCreateUsage } from './token-create-usage';
 import { FeeTool } from '../helpers/calculatePrice';
 import { EstimateCreateCollectionType } from '../types/estimate-create-collection';
-import { getTokenCreateWithFeesUsage } from '../helpers/getTokenCreateWithFeesUsage';
+import { getTokenCreateWithFeesUsage } from '../helpers/get-token-create-with-fees-usage';
 
 export const estimateCreateCollectionInDollars = async ({
   collectionName,
