+++ src/nftSDKFunctions/token-create-usage.ts
@@ -1,5 +1,24 @@
+/*-
+ *
+ * Hedera NFT Utilities
+ *
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 import { TokenCreateUsageProps, TokenCreateUsage } from '../types/estimate-create-collection';
-import { getStringSize } from '../helpers/getStringSize';
+import { getStringSize } from '../helpers/get-string-size';
 import { DEFAULT_EXPIRATION_TIME } from '../utils/constants/fee-tool';
 
 export const tokenCreateUsage = ({
