--- src/utils/hedera/getMirrorNodeUrlForNetwork.ts
@@ -1,3 +1,0 @@
-export const getMirrorNodeUrlForNetwork = (network: string): string => {
-  return `https://${network === 'mainnet' ? 'mainnet-public' : network}.mirrornode.hedera.com/api/v1`;
-};
